Rename derive_variant_test -> derive_godotconvert_test
